requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.5
django-bootstrap4==2.3.1
django-clear-cache==0.3
django-phone-field==1.8.1
django-widget-tweaks==1.4.8
psycopg2-binary==2.8.6
pytz==2020.5
soupsieve==2.1
sqlparse==0.4.1
